Tests/LibWeb/Text/expected/HTML/Window-named-properties-elements.txt

       george  true
true
true
true
true
true
